refactoring get_signal()
