+++ binance_api/interaction_exchange/search_signal_ema.py
@@ -1,25 +1,13 @@
 from loguru import logger
-import pandas as pd
 
-from binance_api.exchange_data.add_dataframe import get_dataframe_pandas, add_dataframe
-from indicators.add_indicators_to_dataframe import add_exponential_moving_average, add_moving_average
+from binance_api.exchange_data.add_dataframe import add_dataframe, adding_dataframe_ema
 
 
 @logger.catch()
-def adding_dataframe_ema(data: list, period_stop: int, period_fast: int, period_slow: int) -> pd.DataFrame:
-    data = get_dataframe_pandas(data)
-    data = add_exponential_moving_average(data, period=period_stop)
-    data = add_exponential_moving_average(data, period=period_fast)
-    data = add_moving_average(data, period=period_slow)
-    return data
-
-
-@logger.catch()
-def add_position(data: pd.DataFrame, period_stop: int, period_fast: int, period_slow: int) -> str | None:
+def add_position(data, period_stop: int, period_fast: int, period_slow: int) -> str | None:
     if data[f'EMA_{period_stop}'].iloc[-1] * 0.998 < data[f'MA_{period_slow}'].iloc[-1] < \
             data[f'EMA_{period_stop}'].iloc[-1] * 1.002:
         return
-
     elif data[f'EMA_{period_fast}'].iloc[-1] > data[f'MA_{period_slow}'].iloc[-1] and \
             data[f'EMA_{period_stop}'].iloc[-1] < data['Open'].iloc[-1]:
         return 'LONG'
@@ -40,23 +28,21 @@ def add_position(data: pd.DataFrame, period_stop: int, period_fast: int, period_
 
 @logger.catch()
 def output_signals_ema(exchange_type: str, symbol: str, time_frame: str, period_stop: int, period_fast: int,
-                       period_slow: int, current_position_last: dict) -> tuple[bool, str] | None:
+                       period_slow: int, current_position_last: dict) -> tuple[bool, str | None]:
     data = add_dataframe(exchange_type, symbol, time_frame, period_stop)
     if isinstance(data, str):
         return False, data
     data = adding_dataframe_ema(data, period_stop, period_fast, period_slow)
     current_position = add_position(data, period_stop, period_fast, period_slow)
     logger.info(
-        f"{current_position_last['position']}/{current_position} price = {data['Open'].iloc[-1]}, "
+        f"Previous position: {current_position_last['position']}/ "
+        f"Current position: {current_position} price: {data['Open'].iloc[-1]}, "
         f"Stop_EMA: {data[f'EMA_{period_stop}'].iloc[-1]}, EMA: {data[f'EMA_{period_fast}'].iloc[-1]}, "
         f"MA: {data[f'MA_{period_slow}'].iloc[-1]}")
-    if not current_position_last['position'] and current_position == "LONG" or current_position == "SHORT":
+    if current_position_last['position'] != current_position:
         current_position_last['position'] = current_position
         return True, current_position
-    if current_position_last['position'] == current_position:
-        return
-    current_position_last['position'] = current_position
-    return True, current_position
+    return False, None
 
 
 if __name__ == '__main__':
